requirements.txt
django-widget-tweaks
Pillow